test: fixing test for cpu metrics on windows
