+++ agent/stats/utils_unix_test.go
@@ -20,6 +20,8 @@ import (
 	"testing"
 
 	docker "github.com/fsouza/go-dockerclient"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func TestDockerStatsToContainerStatsZeroCoresGeneratesError(t *testing.T) {
@@ -36,7 +38,30 @@ func TestDockerStatsToContainerStatsZeroCoresGeneratesError(t *testing.T) {
 	dockerStat := &docker.Stats{}
 	json.Unmarshal([]byte(jsonStat), dockerStat)
 	_, err := dockerStatsToContainerStats(dockerStat)
-	if err == nil {
-		t.Error("Expected error converting container stats with empty PercpuUsage")
-	}
+	assert.Error(t, err, "expected error converting container stats with empty PercpuUsage")
+}
+
+func TestDockerStatsToContainerStatsCpuUsage(t *testing.T) {
+	// doing this with json makes me sad, but is the easiest way to deal with
+	// the inner structs
+
+	// numCores is a global variable in package agent/stats
+	// which denotes the number of cpu cores
+	numCores = 4
+	jsonStat := fmt.Sprintf(`
+		{
+			"cpu_stats":{
+				"cpu_usage":{
+					"percpu_usage":[%d, %d, %d, %d],
+					"total_usage":%d
+				}
+			}
+		}`, 1, 2, 3, 4, 100)
+	dockerStat := &docker.Stats{}
+	json.Unmarshal([]byte(jsonStat), dockerStat)
+	containerStats, err := dockerStatsToContainerStats(dockerStat)
+	assert.NoError(t, err, "converting container stats failed")
+
+	require.NotNil(t, containerStats, "containerStats should not be nil")
+	assert.Equal(t, uint64(25), containerStats.cpuUsage, "unexpected value for cpuUsage", containerStats.cpuUsage)
 }
